+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugs Fixed
 
 - Set the Activity status to `Error` on failed activity source activities.
+- Mark the `Azure.Core.Http.Request` span as failed if the request fails with an exception thrown in the pipeline.
 
 ### Other Changes